Introduce rubocop and replace factory_girl with factory_bot

+++ app/models/bookmark.rb
@@ -1,29 +1,31 @@
+# frozen_string_literal: true
+
 class Bookmark < ApplicationRecord
   belongs_to :site
   has_and_belongs_to_many :tags
 
-  validates :title, :presence => true
-  validates :url, :presence => true
+  validates :title, presence: true
+  validates :url, presence: true
 
   before_validation :set_parent_site
   around_destroy :destroy_orphaned_parent
 
-  scope :tagged_with, -> (tag_name) { joins(:tags).where('tags.name=?', tag_name) }
+  scope :tagged_with, ->(tag_name) { joins(:tags).where('tags.name=?', tag_name) }
 
   def all_tags=(names)
-    self.tags = names.split(",").map do |name|
+    self.tags = names.split(',').map do |name|
       Tag.where(name: name.strip).first_or_create!
     end
   end
 
   def all_tags
-    self.tags.map(&:name).join(", ")
+    tags.map(&:name).join(', ')
   end
 
   private
 
   def set_parent_site
-    url_base = URI.parse(self.url).host
+    url_base = URI.parse(url).host
     if url_base.present?
       url_base = "http://#{url_base}" unless url_base.start_with? 'http'
       self.site = Site.find_or_create_by(url: url_base)
@@ -31,10 +33,8 @@ class Bookmark < ApplicationRecord
   end
 
   def destroy_orphaned_parent
-    parent = self.site
+    parent = site
     yield # executes a DELETE database statement
-    if parent.bookmarks.length == 0
-      parent.destroy
-    end
+    parent.destroy if parent.bookmarks.empty?
   end
 end
